Декмопозиция Login-page, создание слайса auth-slice
